Part 1 — Build ETL Pipelines/pyproject.toml
[tool.poetry]
name = "gd-nasa-pipeline"
version = "0.1.0"
description = ""
authors = ["Nacima Yahiatene <nacima.yahiatene@gnomondigital.com>"]
packages = [
    { include = "gd_nasa_pipeline", from = "src" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
requests = "^2.31.0"
botocore = "^1.34.2"
pytz = "^2023.3.post1"
netcdf4 = "^1.6.5"
xarray = "^2023.12.0"
numpy = "^1.26.2"
python-dotenv = "^1.0.0"
sqlalchemy = ">=1.4.28,<2.0" 
psycopg2-binary = "^2.9.9"
click = "^8.1.7"
boto3 = "^1.34.15"
pytest = "^7.4.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

[tool.poetry.scripts]
app = 'gd_nasa_pipeline.main:run'

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/ 
'''

[tool.isort]
line_length = 79
lines_after_imports = 2